+++ apps/docs/src/components/colors/Colors.tsx
@@ -1,5 +1,5 @@
 import * as typographyStyles from "gu-ds-css/output/typography.module.css";
-import tokens from "gu-ds-tokens/output/tokens.json";
+import tokens from "guds-tokens/output/tokens.json";
 import type { FunctionComponent } from "react";
 import { Page } from "../page/Page";
 import { RichText } from "../richText/RichText";